Make generic implementations more poly-kinded
`gftraverse`, `gfzipWith`, and `gfrepeat` all monomorphize the kinds
of type variables (e.g., `f` and `g`) to `Type -> Type`, but this can
be generalized to `k -> Type` quite straightforwardly. I've modified
one of the examples to ensure that this generalization actually does
happen.

In order to preserve the existing order of type applications in these
functions, I opted not to explicitly quantify the kind variable `k`.